Flesh out the AIR example scripting.
Rename the ‘download.sh’ to ‘setup.sh’, because that’s what it does.

Stick a face on Smiley, and do something with it.

Move the AIR build outputs down one level, because the script was
including builds themselves recursively.

Ignore build ouputs, because it’s a lot of ‘always different’, that can
just be built.