Carrying on switching setup to ant
Move the scary setup download/unzip script from shell script to AIR
build script.  Now 3x as scary, but at least it *should* be an all in
one tool, rather than needing GNU bits that may or may not be installed.

Setup the ‘readme’ to reflect that change, though it is only for
Windows, for the moment.