Add publish folder to excludes, automatically.
Because I have no need to keep syncing a bunch of binaries that this is
supposed to be making.

Also, see some interesting/odd behavior on ‘D:\’ drive, under Windows.
Looks like Cocos doesn’t like that.  I will touch that tomorrow, or add
a note to the setup documents, or whatever.  I’m too sleepy to function.